Refactored Random Language Model into its own module; code cleanup in icon task
